+++ src/hooks/index.js
@@ -1,7 +1,4 @@
 export * from './useAnimation';
-export * from './useIsLoading';
-export * from './useWindowSize';
-export * from './useVars';
 export * from './useColors';
 export * from './useContract';
 export * from './useLogger';
@@ -17,3 +14,4 @@ export * from './useTokenConstant';
 export * from './useConstants';
 export * from './useTranslation';
 export * from './useTracking';
+export * from './useBuyProviders';