+++ src/routes/Login/Login.js
@@ -1,13 +1,13 @@
 import React, {useEffect, useRef, useState} from 'react';
 
-import {LoginErrorMessage, WalletLogin} from '../../components/UI';
+import {MultiChoiceMenu} from '../../components/UI';
 import {
+  ActionType,
   ChainInfo,
   LoginErrorType,
   NetworkType,
   WalletErrorType,
-  WalletStatus,
-  WalletType
+  WalletStatus
 } from '../../enums';
 import {
   useEnvs,
@@ -16,9 +16,9 @@ import {
   useWalletHandlerProvider
 } from '../../hooks';
 import {useHideModal, useProgressModal} from '../../providers/ModalProvider';
-import {useIsL1, useIsL2} from '../../providers/TransferProvider';
+import {useIsL1, useIsL2, useTransfer} from '../../providers/TransferProvider';
 import {useWallets} from '../../providers/WalletsProvider';
-import utils from '../../utils';
+import {evaluate, isChrome} from '../../utils';
 import styles from './Login.module.scss';
 
 const MODAL_TIMEOUT_DURATION = 2000;
@@ -35,21 +35,21 @@ export const Login = () => {
   } = useLoginTranslation();
   const [trackLoginScreen, trackDownloadClick, trackWalletClick, trackLoginError] =
     useLoginTracking();
-  const {autoConnect, supportedChainId} = useEnvs();
+  const {autoConnect, supportedL1ChainId} = useEnvs();
   const [selectedWalletName, setSelectedWalletName] = useState('');
   const [error, setError] = useState(null);
-  const [walletType, setWalletType] = useState(WalletType.L1);
+  const [, swapToL1] = useIsL1();
+  const [, swapToL2] = useIsL2();
+  const {action} = useTransfer();
+  const {status, error: walletError, connectWallet, isConnected} = useWallets();
   const modalTimeoutId = useRef(null);
   const hideModal = useHideModal();
   const showProgressModal = useProgressModal();
-  const getWalletHandlers = useWalletHandlerProvider();
-  const {status, error: walletError, connectWallet, isConnected} = useWallets();
-  const [, swapToL1] = useIsL1();
-  const [, swapToL2] = useIsL2();
+  const walletHandlers = useWalletHandlerProvider(action);
 
   useEffect(() => {
     trackLoginScreen();
-    if (!utils.browser.isChrome()) {
+    if (!isChrome()) {
       setError({type: LoginErrorType.UNSUPPORTED_BROWSER, message: unsupportedBrowserTxt});
     }
     return () => swapToL1();
@@ -60,17 +60,18 @@ export const Login = () => {
     if (error) {
       trackLoginError(error);
     } else if (!error && autoConnect) {
-      const handlers = getWalletHandlers(walletType);
-      if (handlers.length > 0) {
-        timeoutId = setTimeout(() => onWalletConnect(handlers[0]), AUTO_CONNECT_TIMEOUT_DURATION);
+      if (walletHandlers.length > 0) {
+        timeoutId = setTimeout(
+          () => onWalletConnect(walletHandlers[0]),
+          AUTO_CONNECT_TIMEOUT_DURATION
+        );
       }
     }
     return () => clearTimeout(timeoutId);
-  }, [error, walletType, getWalletHandlers]);
+  }, [error, walletHandlers]);
 
   useEffect(() => {
     if (isConnected) {
-      setWalletType(WalletType.L2);
       swapToL2();
     }
   }, [isConnected]);
@@ -113,10 +114,9 @@ export const Login = () => {
   };
 
   const onDownloadClick = () => {
-    trackDownloadClick(walletType);
-    const handlers = getWalletHandlers(walletType);
-    if (handlers.length > 0) {
-      return handlers[0].install();
+    trackDownloadClick();
+    if (walletHandlers.length > 0) {
+      return walletHandlers[0].install();
     }
   };
 
@@ -124,8 +124,8 @@ export const Login = () => {
     if (error.name === WalletErrorType.CHAIN_UNSUPPORTED_ERROR) {
       setError({
         type: LoginErrorType.UNSUPPORTED_CHAIN_ID,
-        message: utils.object.evaluate(unsupportedChainIdTxt, {
-          chainName: ChainInfo.L1[supportedChainId].NAME
+        message: evaluate(unsupportedChainIdTxt, {
+          chainName: ChainInfo.L1[supportedL1ChainId].NAME
         })
       });
     }
@@ -134,10 +134,7 @@ export const Login = () => {
   const maybeShowModal = () => {
     maybeHideModal();
     modalTimeoutId.current = setTimeout(() => {
-      showProgressModal(
-        selectedWalletName,
-        utils.object.evaluate(modalTxt, {walletName: selectedWalletName})
-      );
+      showProgressModal(selectedWalletName, evaluate(modalTxt, {walletName: selectedWalletName}));
     }, MODAL_TIMEOUT_DURATION);
   };
 
@@ -149,40 +146,38 @@ export const Login = () => {
     hideModal();
   };
 
-  const renderLoginWallets = () => {
-    return getWalletHandlers(walletType).map(walletHandler => {
+  const mapLoginWalletsToChoices = () => {
+    return walletHandlers.map(walletHandler => {
       const {
         config: {id, description, name, logoPath}
       } = walletHandler;
-      return (
-        <WalletLogin
-          key={id}
-          description={description}
-          isDisabled={!utils.browser.isChrome()}
-          logoPath={logoPath}
-          name={name}
-          onClick={() => onWalletConnect(walletHandler)}
-        />
-      );
+      return {
+        id,
+        description,
+        isDisabled: !isChrome(),
+        logoPath,
+        name,
+        onClick: () => onWalletConnect(walletHandler)
+      };
     });
   };
 
   return (
-    <div className={utils.object.toClasses(styles.login, 'center')}>
-      <div className={styles.content}>
-        <div className={styles.title}>{titleTxt}</div>
-        <p>
-          {utils.object.evaluate(subtitleTxt, {
-            networkName: walletType === WalletType.L1 ? NetworkType.L1.name : NetworkType.L2.name
-          })}
-        </p>
-        <div className={styles.container}>{renderLoginWallets()}</div>
-        {error && <LoginErrorMessage message={error.message} />}
-      </div>
-      <div className={styles.separator} />
-      <div className={styles.download}>
-        {downloadTxt[0]} <span onClick={onDownloadClick}>{downloadTxt[1]}</span>
-      </div>
+    <div className={styles.login}>
+      <MultiChoiceMenu
+        choices={mapLoginWalletsToChoices()}
+        description={evaluate(subtitleTxt, {
+          networkName:
+            action === ActionType.TRANSFER_TO_L2 ? NetworkType.L1.name : NetworkType.L2.name
+        })}
+        error={error}
+        footer={
+          <div className={styles.download}>
+            {downloadTxt[0]} <span onClick={onDownloadClick}>{downloadTxt[1]}</span>
+          </div>
+        }
+        title={titleTxt}
+      />
     </div>
   );
 };