+++ src/utils/blockchain/index.js
@@ -1,2 +1,0 @@
-export * as ethereum from './ethereum';
-export * as starknet from './starknet';